monitoring-agent/requirements.txt
requests>=2.31.0
cryptography>=41.0.0
psutil>=5.9.0
gitpython>=3.1.40
python-dotenv>=1.0.0
watchdog>=3.0.0
